Aggregate indexeddb process
